RAG_EASY/requirements.txt
gradio
langchain
langchain-openai
langchain-community
python-dotenv
faiss-cpu
openai
